drive bugfixes & replaced exp2 with pow 2
Changes made:
  * (#) Enabled DMN and turning multiplier to actually be used
  * (~) updated zig.mk
  * (#) switched @exp2 with std.math.pow(x, 2)

Recently I discovered that both DMN and the turning multiplier both
weren't being properly applied to the user arcade drive.
That's because, I was using the raw x and y, instead of the nx and
ny, of which I was applying the modifications to.

Also, replaced all the instances of @exp2 with std.math.pow(x, 2) as
for certain values, the @exp2 returned different values than
std.math.pow(x, 2) and messed up a bunch of my calculations.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>