Revert "removed tower motor A"
Changes Made:
  * (^) reverts commit a024abf52ae618ff34d11a439992cd7db14b132f.
  * (~) updated port numbers for A & B
  * (~) changed scoring high and middle behaviours
  * (#) kept the %d -> %lu time format bugfix

Recently we added another motor to the tower to support scoring high
goals.

So I've just repurposed the old A tower motor to be this new moter.
As the motor is under the old B motor, I've shift the B motor to be
called A, and then made the new motor to be B.

The new B motor, always spins in the same direction as the old B motor
(the now A motor).

I've also kept the same %d -> %lu bugfix.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>