compeition commit
Changes made:
  * (~) changed pid values to be P only and small (no tuning yet)
  * (~) modified and tweaked daniel's magic number & the turning multiplier to feel nicer
  * (~) commented out Spaghett0's auton path, wrote a last ditch effort, still didnt work...
  * (#) removed daniel's magic number from auton arcade, and made it functional in driver arcade
  * (~) swapped the controls for scoring goals by driver request
  * (~) added 2s delay to tuning runs

These were blazing fast changes made to the robot during our first comp,
our only real time we got to test.

We tried to tune the pids, but ultimately didn't have time.
It didn't matter anyways cuz autonomous, despite existing, just didn't
run, and seemed to be skipped for some reason.

I'm still not quite sure the reason as to why.

Also, from the feedback from the drivers, the exponential was too
exponential and the turning multiplier was too high.
So we turned both those down to meet those needs.
After doing some research, we're probably going to ditch the exponential
controls as they are akin to mouse acceleration for games, which is
undesireable due to it's lack of intuitivity and imprecise-ness.

Also, after double checking the arcade drive functions, I realized that
the both the normal and driver arcade both generated new Xs and Ys and
then just used the old X and Y values (ignoring the new values).
For the normal driver arcade, we simply just removed the generation of
DMN.
And for the normal/auton arcade drive, I just made it so that the
function used the new X & Y values.

But that doesn't matter anyways as we're dropping DMN anyways, so we're
likely to unify the arcade drive functions (as that was for keeping the
now defunct curve).

Also, the tower controls are now swapped, as that feels more intiutive
for our driver.

And the tuner runtimes now also come with a 2.5s timer, as, yes this can
be done with a field controller, but getting and plugging in a field
controller in extremely inconvenient.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>