no more daniel's magic number :c
Changes Made:
  * (-) removed daniel's magic number from the drive code
  * (-) removed the daniel's magic number build option
  * (~) replaced the userArcadeDrive function with normal arcadeDrive

After doing extensive research, and personal driver testing, we have
concluded that linear drive is far surperior to exponential drive,
and all benefits provided by it, can be easily substituted.

Online, almost every source you find discusses how exponential drive is
unintuitive and hard to use, with it's only benefit, making a game harder
; which is the last thing we want.
*It was also compared to mouse acceleration, which is notorious for
being counter-productive*

And, through analyzing the recorded videos of our matches, you can see
our driver have to start and stop and then turn, due to the exponential
ramp up of the forwards movement of the joystick.
And how, there's essentially a giant deadzone on the joystick where the
robot doesn't move at all and then it ramps up to full speed almost
instantly.

And, a pattern was noticed, that as we adjusted the DMN curve to be more
linear, the drivers perfered it better.

Also, the main benefit of having an exponential drive, in fact, the only
reason why we even considered it in the first place, was because of us
previously using a voltage-based drive.
We wanted to still both reduce the speed of turning and movement,
without sacrificing torque.
But now, since we use velocity drive, the robot will automatically
increase torque to maintain a set velocity, redering the benefits of
exponential drive irrelevant.

Also, due to the removal of DMN, the userArcadeDrive function is no
longer necessary, and the turning multiplier can just be applied
individually for each of the arcade implementations.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>