movement multipliers and locking speed boosts
Changes made:
  * (+) added movement multiplier
  * (~) moved toggle arcade into it's own function
  * (+) added a speed boost when locking to movement
  * (+) added a speed reduction when locking to rotation

Recently, something that was noticed during the compeition, was that,
turning was far too slow compared to movement, so our driver would have
to stop, and then turn and then move again- defeating the purpose of
having arcade-drive in the first place.

Now this problem isn't present in split arcade nor tank drive, but they
both don't feel intuitive, which matters most.

Now, this problem can't simply be solved by bumping the turning speed up
as, as it stands, the turning speed is already farrr too fast already.
So, our only option is to reduce the movement speed.

However, I still wish to be able to have the robot zip across the field
at mach 5.
So, a solution I've thought of, that extends toggle arcade, is that,
since you don't want the robot to turn anyways when the robot is
travelling incredibly fast, so what if, the robot speed is boosted
when the robot is locked to vertical movement.

And also, the robot turning speed is reduced when it's locked to
rotating as you only tend to do that, when you want to do a precise turn.

Also toggle arcade drive is complicated enough to deserve it's own
function.

I might also experiment with having a turning speed larger than a
movement speed, as hitting the toggles flips that, and that makes the
toggles made more sense.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>