drive: combined the movement & turning multipliers
Changes made:
  * (~) combined the movement & turning multipliers into one number

I was thinking, and having the movement speed be faster than the turning
speed was an artifact of wanting to both have precision (a low floor)
and also the ability to have high speeds (a high ceiling).

But since that's now solved by the toggle arcade speed boost toggles,
I can make the drive feel WAYYY more intuitive by making the movement
and turning speeds THE SAME, as, that just feels better and makes more
intuitive sense in my opinion.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>