GIANT COMMIT: 'fully working' custom pure pursuit
Changes made:
  * (~) .gitignore: added ignore clause for generated CSV data
  * (#) logging.zig: added correct units and conversions (mm/ms instead of mm/s) (mm/ms = m/s)
  * (%) std.math.pow: replaced all instances of 'power by 2' with it multiplied by itself (optimization)
  * (~) odom.zig: added actual units to doc comments
  * (#) odom.zig: usages of undoLatOffset and also calculating rot_vel used subtraction for angle difference instead of the more correct minimalAngleDiff
  * (#) swapped all vectors to use left-handed y-based directions (instead of the standard maths right-handed x-based)
  * (#) zig.mk: removed DMN from the project build command

  * (+) pure_pursuit.zig: ADDED MY OWN CUSTOM PURE PURSUIT ALGORITHM!!!
    * (+) added path segment picking function (alongside tests for it)
    * (+) added goal interpolation from path segment, robot coord & search radius function
      * (+) also added a basic test for it that seems right
      * (#) replaced exp2 with the correct x*x
    * (+) added circle/arc fitting and following algorithm (thanks mr bui)
      * (~) also, included on spot rotation clause for angle err (theta) > 90
      * (#) added theta_sign to keep the correct signs for the left & right sides
      * (+) added a simple 'test' for arc following (in the case the robot's going forwards)
  * (+) ADDED A FULL ON FORWARDS DIFFERENTIAL DRIVE KINEMATICS SIMULATION TO TEST PURE PURSUIT!!!
    * (#) added a cycle limit in case pure-pursuit is bugged and the robot never reaches it's goal

Since this program generates .csv data files as logs for the robot, and
also even within 'unit' tests that the program runs (forwards kinematics
simulation) these csv files are created.
These files themselves are just raw data used for analysis or testing
and shouldn't be included in the repository itself; only any insights
learnt or gained from them.

Also, looking through the velocity logs generated by the robot, I noticed
that the velocity values reported by the robot we ridiculously low,
like 0.01mm/s which just cannot be valid under any context.
So I did some digging, and found out that all the calculations were
actually done with dt being in milliseconds, and instead of changing the
calculation itself, I've chosen to just do a conversion before we log it.
The neat thing is, however, that mm/ms cleanly translates to m/s without
needing any conversions.
I've also added these units to the doc comments of odom for
disambiguation.

Also, I've made a simple optimization for the instances in which I find
a value to the power of 2, by simply multiplying it with itself.

Some more general bugfixes that I've caught and fixed, are that before,
for functions like calculating the rotational velocity and undoing the
lateral odom sensor offset used the subtraction difference between the
angles instead of the more correct minimalAngleDiff (as otherwise it may
take the longer route around the circle).
Also removed the now non-existant daniel's magic number from the zig
makefile.

Also, I've chosen to use left-handed y axis aligned angles for the
vectors I use in my code instead of the standard right-handed x axis
aligned that's used in maths.
I've chosen to do this as if I were to use the standard angle arrangement
I'd have to convert it *every* time I used it as there is no instance
where I don't need it in the form left handed y-axis form as that's the
form the robot yaw is in.

Now, for the meat of this commit, PURE PURSUITT!!!
(and also the simulation I wrote to test it)

I've implemented my own pure pursuit algorithm, and I could go into depth
about it's implementation details, though I've already written it down
some other place, and it's not worth writing that much for a git commit
message.

The main overview is, you have an array of path points that you want the
robot to go to, the robot has a radius around it called the search_radius
and that's the circle upon which it will pick 'goal points' that it'll
head towards.
The first function picks the segment (start and end points) of the path
that the robot is following.
Then, once it has those two points, it can first cast the robot's
position onto the path, and then figure out the extension provided by
the search radius from the search radius and how much the robot's
strayed from the path, and with that extension you can calculate the
'goal point'.
Then, the final function, finds the radius of a circle that connects the
robot position (with the robot's yaw tangent to the circle), and the goal
point and returns the ratio of the left and right wheels to get the robot
to drive in that arc.

Some noteable things and bugfixes that might not be included in
explanations of my algorithm (discovered through testing).

Previously in my initial implementation of goal point interpolation
mistakenly used the zig builtin @exp2 which I thought was for squaring
values, when instead it was 2**x, intead of x**2 (GIANT MISTAKE), so
instead I've switched to just having the value multiply itself.

Also, a notable thing about the circle/arc fitting & following function's
implementation, is that when the angle error (theta) is (abs) larger than
90 degrees, the robot doesn't attempt to fit a circle, and instead just
turns towards the end point on the spot.
The reasoning behind this, is first to fix a bug where, if theta is
180, the circle the robot fits to get the robot to reach it's goal point,
is infinitely large, and it's taking the long way (will never reach).
As, 90 degrees, is the point where, the path the robot takes around it's
arc, is exactly half the circumference of the circle, and any error
larger than that, would result in the robot travelling the longer way
around the circle.
Which isn't reasonable in my eyes, so I've clamped it to 90.

Also, I've had to include a theta_sign to keep the correct signs for the
left and right sides so the robot turns the right direction.

Now, all of this testing was done on a simulation.
What I'd done last year, was make a full on bevy game for these sorts of
simulations, and have it play out in real time.
This, while cool, is not very useful for debugging, as I want to see
the path the robot actually takes.
Now, I could add some sort of path tracing, and write it in a game engine
in zig- ORRRRR I could do what I did with logs.
In that, I just simulated the values robot would log (in CSV) for it's
odom position through forward differential drive kinematics and graph it
like normal through an X Y scatter plot, which solves all my problems
and looks really cool with minimal effort.

Also, either the robot reaches the goal (is within the precision limit
defined in auton) or it takes too long and hits the cycle limit I've
introduced, where the test just panics and gives a nice message instead
of getting stuck in an infinite loop.

note to future kalscium:
  if you wish to better understand the pure
  pursuit algorithm that I've written, then search through your oats with
  the tag 'SMTIFR'.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>