added basic speed controller and tunable parameters
Changes Made:
  * (+) pure_pursuit.zig: added a speed controller function
  * (+) pure_pursuit.zig: added a centralized struct for pure pursuit parameters & tuning advice
  * (+) pure_pursuit.zig: added robot coord & yaw prediction function
  * (+) pure_pursuit.zig: added tests for the prediction func

I've implemented a pretty simple velocity controller for the robot that
I've thought of.

The three major components of this velocity controller are:
  * kp             -> distance errors
  * turn_180_speed -> yaw errors
What this means, is that the robot will slow down the smaller the
distance error (depending on kp), and also slow down the larger the
yaw error is (depending on turn_180_speed).
This makes intuitive sense, as, you want to slow down as the robot
approaches it's goal, and you also want the robot to slow down before
turns (when there is a large yaw error).

What kp actually is, is just the speed the robot will travel at when
the goal point is outside or on the search radius.
So, the distance will never actually be larger than the search radius,
and will only decrease once the robot is reaching the true path end point.

Also, what the turn_180_speed parameter actually is, is the speed the
robot will turn at when the yaw error is 180 degrees, as that's the
largest turn the robot *can* make.
You just specify the lower bound for the robot speed when turning, and
the rest is lerped.

Now the proportional and turning portions are multiplied together as they
are both within the 0..1 range, and are meant to be used like multipliers
and therefore can be combined like so.

I've also included a centralized parameters struct for all the parameters
for pure pursuit so all the parameters can be conveniently tuned and
stored in one location.
Also will be critical for passing it into future convenience functions
like move_to or something along the lines of that.

Also, there is now a function that predicts the robot's next position
based upon a lookahead window and the robot's current velocities.
So far it only operates on velocity and not acceleration, as that would
introduce far too much noise where the robot reacts to it's own reaction
and fights itself.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>