bugfixes & added pure pursuit state machine func
Changes made:
  * (+) autonomous.zig: added 'tuned' pure pursuit parameters to autonomous
  * pure_pursuit.zig:
    * (+) added convenience state machine for following path
    * (~) changed follow arc to return a vector instead of a struct
    * (#) fixed speed controller to take in a coord for coord instead of f64
    * (#) added missing type annotatations for calMag function calls
    * (#) changed 180.0 degree usage to be it's radian counterpart (pi)
    * (~) extended the forwards kinematic sim to include speed controller calculations

First, I've added 'tuned' pure pursuit parameters to autonomous.zig, that
will be both used for autonomous, and also within the convenience
state-machine function
(the purpose of which is to make it as frictionless to follow paths).

The main change/addition introduced by this commit, is the convenience
state machine for pure pursuit, which like for PID or major minor state
machines, in that they abstract away the actual path following into one
function call which is fine as it doesn't actually introduce any
additional functionality beyond that convenience.

It does this by directly driving robot in a tight feedback loop using
parameters defined in autonomous.zig.

The only notable thing about it's implementation, is the
'driving in reverse' functionality.
Where, when enabled, the robot will follow the specified path as if the
back of the robot was the front.
How this is implemented, is that, when enabled, the odom state is copied
and changed to appear how as it would, if the robot's orientation was
rotated 180 degrees.
And, any 'instructions' generated for the robot in this reversed mode
are swapped and inverted to fit the actual orientation of the robot.
This is done this way, mainly for convenience reasons, so that all the
calculations are the same for forwards and backwards, only the input and
output data has a little bit of conversions.
Also, it's only the pure pursuit that's done in 'reverse', odom and
everything is still calculated in the normal orientation, for, again,
simplicity reasons.

Also, I've changed followArc to output a vector instead of a struct for
performance reasons, so I can perform operations to both left and right
drives in parrallel (through SIMD).

Also, in many places in the code, I forget that the code deals with
angles in radians, and not degrees, and so that introduced many bugs.
Anyhow every now and again, just grep for 180 and 90 and replace them
with their respective radian values.

Also, even though the forwards kinematics simulation doesn't have physics
and therefore the speed of the robot effects litterally nothing, I've
still included the option to graph what the speed output *would* look
like from the speed controller, to aid in debugging (which it has).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>