cycle simplification
Changes made:
  * (~) opcontrol.zig: replaced existing timing-based system with a cycle-based system
  * (~) opcontrol.zig: renamed tick_delay to cycle_delay

I've replaced the current timing-based system with multiple states in
favour of a more sychronized and simpler cycle system, where we ensure
that each cycle runs, one after the other, so that regardless of timing
the actions will happen after a specified amount of cycles.

This is far simpler, faster and easier to manage than the old system.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>