debug mode: added debug opcontrol mode for tuning
Changes made:
  * (+) build.zig.zon: added debug_mode build option
  * (~) autonomous.zig: renamed tick_delay to cycle_delay
  * (+) debug_mode.zig: added a debug opcontrol mode for tuning, testing, planning and showcasing for auton

It's 6:25 AM right now.

I've implemented the first basic version of this debug mode that I had an
idea for.

The basic controls are as follows:
  A -> pushes the current location to the recorded auton path stack
  B -> pops the last-pushed odom location from the path stack
  X -> starts or stops the auton (toggle)
  Y -> restarts the robot from the start of the recorded auton path
  LEFT -> moves the order of magnitude increment decimal place to the left
  RIGHT -> moves the order of magnitude increment decimal place to the right
  UP -> cycles through parameters to tune
  DOWN -> cycles through parameters to tune
  L2 -> doubles the current parameter value
  L1 -> halves the current parameter value
  R2 -> increments by the order of magnitude
  R1 -> decrements by the order of magnitude

Also, when the robot is paused, you can move it around like normal driver
control, except it's tank drive, and incredibly slowed down.

ALSO, when Y is hit, and everything (except for the path stack & parameters)
is reset, the robot also logs down all the important information of things
that might've changed.
Like the current state of the path point stack, or the current state of
the tuned parameters, or the port buffer (for port disconnects).

Note, that none of this code is tested so far, so expect none of it to
work first try.
We have testing tomorrow (today), and we'll iron out all the bugs and
also tuned this pure pursuit.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>