pure pursuit: (after) IRL testing and tuning
Changes made:
  * autonomous.zig:
    * (~) (arbitrary) changed auton's distance precision
    * (~) tuned yaw pid (and disabled the integral & derivative terms)
    * (~) somewhat tuned pure pursuit parameters and added some helpful comments
    * (#) found and fixed the reason why auton wasn't working at comp
  * debug_mode.zig:
    * (#) made the robot start with auton paused
    * (#) added round-robin rumble queuing to respect controller 50ms update delay
    * (+) added the lookahead-window prediction to pure pursuit runs
    * (#) 'fixed' the log file outputs (style/format)
    * (#) added resetting of the odom sensors aswell when Y is hit
    * (#) accidentally binded parameter increment/decrementing to LEFT/RIGHT instead of R1/R2
    * (~) expanded the round-robin controller updates to be 20 cycles (200ms) for rumble support
  * (~) main.zig: added a debug hello world print for the initialize function
  * odom.zig:
    * (~) odom.zig: made rotation ports public for debug_mode Y resets
    * (~) odom.zig: added negation to counteract physical reversal of vertical rotation sensor
  * (#) pure_pursuit.zig: added ACTUAL robot width (290mm instead of 4mm :I) (not width but distance between middle of wheels)
  * (#) tuner.zig: updated the name of tick_delay to cycle_delay

So, yesterday (Friday, 19th of July 2024), I finally had physical access
to the robot for ~4 hours to test out, tune and fix both my driver and
autonomous code.
We didn't actually have access to the regulation vex anti-static tiles
at the time, so we haven't done any actual proper tuning, just general
testing until we have access to the field again.

First, I changed the auton required distance precision from 5mm to 10mm
which is more reasonable for the robot to achieve, this was just for
testing purposes, and I didn't actually get to try out a multitude of
values (for tuning) yet.
10mm works fine, but 5mm would be nicer, if the margin of error isn't
too small.

Next, I also 'tuned' the rotation/yaw PID as I plan to use pure-pursuit
to get to the location where I want the robot to be, and then use the
yaw PID to get it oriented the way I want it to be.
The PID so far, is beautifully clean, with a nice speed and stop and I
probably won't touch it anymore.
Turns out, just the proportional term is actually enough, as we're
setting the velocities of the motors (not the voltages) and so the
built-in motor PIDs do most of the heavy lifting for us
(no steady-state error).

I've also tuned pure pursuit a little, tried out a bunch of different
search radiuses, lookahead windows and turn speeds.
I didn't get to properly tune everything since we didn't have the correct
tiles anyways.
However, I did find that the previous search radius was far too small,
and that increasing it to 240 reduced the osccilations whilst keeping
enough accuracy, but we could keep going and see if we could remove all
osccilations by increasing to like 300 or something similar.
Also, we found that, for proportional speeds, 0.3 was too slow and didn't
have a large accuracy increase to warrant such a slowdown.
We also found 0.6 was WAYY to fast to be reasonable for auton, so you
could try 0.5 and see if you like it or not.

Also, after testing turning speeds a little bit... *turns* out they don't
work AT ALL.
For some reason, it's inverted, where when you decrease the parameter
(instead of increasing) it actually sped up the turning which is the
opposite effect of what we were looking for.
And, after increasing the parameter, which *did* slow down the robot a
little at turns, I've discovered that it doesn't really help all that
much and is INCREDIBLY jank, as the turn multipler sudden kicks into
effect once the path point end is within search radius, and I just prefer
a constant movement speed anyways.
So I've set it to 1 to 'disable' it for now, just like with the yaw
PID parameters.
Might remove it later.

Also, the lookahead window I haven't properly tested yet, and I don't
even know if it's actually making the robot more jank or doing anything
at all.
We'll need more testing, but from analyzing recordings, I've seen that
the robot prematurely stops, which could indicate that the current value
is far too large.

ALSO, I discovered why auton "wasn't running" during our competition.
The reason?, well after some testing, we discovered it in fact *was*
running just that, it didn't move.
And after more testing, we found the issue- the wait function.
So, during the competition, our auton was just a rushed last-ditch have
the robot run at full speed, wait 200ms then stop.
And turns out, I had the wait function check if the user program has
been running for at least 200ms or not (which of course it has) instead
of if 200ms has elapsed or not.
So it just skipped the wait and turned off the motors before we could
react.

Also, turns out, the rumbles of the controller weren't actually going
through most of the time, as they would collide with the controller
screen updates and be lost.
(as the robot controller can only receive 1 update every 50ms).
So, to solve this, I just made a simple rumble state that would be
updated by the code, as I only want the latest rumble if there are
multiple- in essence making a sort of queue.
And, by incorperating the rumble updates into the round-robin system
I currently have set up, the rumbles are neatly queued in with the rest
of the controller updates.

Also, I'd forgotten to add the lookahead window predicitions to my debug
mode runtime, so I just did that.

There were also some formatting mistakes (with spaces) in the debug mode
log so that was solved aswell.

And, the robot position was a little off each and every time we reset
it with Y in debug mode, and so, after some thinking, I think it's
due to the rotation sensors not being reset
(actually does that make any sense?).
I don't know, but that was my reasoning at the time.

Also, to appease my paranoia that certain functions of the robot code
weren't running, I added another few hello worlds.

Also, we discovered the robot's vertical rotation sensor was actually
reversed (to the motors) so when the robot moved forwards the sensor
would read that as going backwards.
To fix this, we just added another line in the odom init stage to set
the reversed flag for that sensor.

Also, initially, the robot's movement with pure pursuit was really jank
as it would start and stop when trying to turn instead of one smooth
movement.
After giving up on fixing it, I started inputting robot measurements into
the code and discovered that the reason for this, was that, for testing
purposes, I'd initially set the robot width to be 4mm instead of it's
actual 290mm.
And, due to that, the robot would due much smaller turn ratios than what
they should've been, and once this was corrected the robot started moving
incredibly smoothly and nice.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>