removed pure pursuit 180 turn speed multiplier
Changes made:
  * (-) pure_pursuit.zig: removed speed controller & 180 turn speed param (also from the simulation)
  * (-) debug_mode.zig: removed speed controller from debug mode
  * (-) autonomous.zig: removed 180 turn speed mul from params

I've chosen to remove the 180 degree turn speed multiplier, for reasons
I've stated before.

First, because it just plainly doesn't work, and second, during testing,
I found that the best value for it, was 1, in that disabling it produced
the smoothest result, without actually having any detrimental effect
on the accuracy.

I found the speed multipler to be incredibly jank, as the robot would
suddenly decelerate once the path end was within search radius.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>