dynamic pursuit: implemented an experimental algo
Changes made:
  * (~) autonomous.zig: converted pure pursuit to dynamic pursuit parameters
  * (~) debug_mode.zig: updated to support dynamic pursuit parameters
  * (+) pure_pursuit.zig: implemented dynamic pursuit

In this commit, I've implemented an idea for an extension to the pure
pursuit algorithm that I'll call 'dynamic pursuit', due to the
dynamically chosen search radius.

The main issue that this extension to the pure pursuit algorithm tries
to solve, is the trade-off between having a large search radius and a
small one.
Where, if the search radius is too large, then it'll round the turns and
corners of the path and be too inaccurate, but conversely, if it's too
small, then the robot will osccilate and not be smooth enough.
Instead of settling for a medicore compromise between the two that's not
close to the best in either aspect.
What if you could have the best of both worlds?

The idea for dynamic pursuit, is that, you have two bounds for the search
radius; a ceiling and a floor.
And, the search radius is computed to be the distance from the path
segment start.
So that naturally, the robot would have a tighter turn (I think).

This commit, is just a commit for the sake of it, as I've already thought
of a far superior solution (dual pursuit), and this is just due to me
not wanting to delete all these changes.

I mean, this might work, we'll see tomorrow when I test this.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>