Does "removed pure pursuit 180 turn speed multiplier" properly
This reverts commit 5a323e2308c36cd32f59e8aa1cbc08b3e1cb4d62.

Changes made:
  * (#) removes the turning pure pursuit 180 turn speed multiplier WITHOUT removing the speed controller
  * (~) replaces the old auton routine to include pure pursuit & PID so the functions get proper error checking
  * (#) renamed tick_delay to cycle_delay in pure pursuit follow arc function

So, the way I'd removed the pure pursuit 180 turn speed multiplier was
rather short-sighted and mistaken, as I'd removed the speed controller
entirely, which means that even the proportional term isn't applied
and the robot then travels at FULL SPEED.

I've updated it, so that it's just removed the 180 turn portion of the
speed controller.

Also, I've replaced the old auton routine with a new pure pursuit & PID
routine, where the robot just moves forwards for a metre, and then points
forwards.
This is mainly just so I get proper error handling and linting of the
pure pursuit followPath function and the PID rotate functions as zig
lazy compiles.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>