auton: made a simple mock left-side auton for comp
Changes made:
  * autonomous.zig:
    * (~) eyeball tuned the pure pursuit parameters
    * (~) moved the auton routine into it's own left-side function
    * (+) added the 'simple' auton routine I'd described (no little-will)
  * (#) pid.zig: added drive stopping when precision is met
  * (#) pure_pursuit.zig: added drive stopping when precision is met

First-- from clips I'd recorded at the builder's house and also from
memory, I've eyeballed some safe adjustments to the pure_pursuit
parameters so that it should work fine tomorrow when we test.
From the clips and memory, three main things are apparent: the robot is
osccilating a little, it's a bit slow, and it's stopping prematurely.

Those symptoms neatly tie into each parameter so it's pretty easy to
guess a more optimal value.

Also, the reason why I chose 240mm for the search radius initially was
because 280 was too inaccurate, and 200 osccilated way too much.
But since 240 is still osccilating a tiny amount, I think 260 is more
appropriate.

I'm tempted to just have a lookahead window of zero because of the
pre-mature stopping. But perhaps it's actually doing something, and 20
to 0 feels like too large of a change.

Also, a giant bug I caught before testing was with the PID and
pure-pursuit convenience functions.
Where, even though the controller is broken from, the last drive
instructions aren't actually reset, so the robot may still move when
it's supposed to have reached it's precision limit.
So I've just added a few patches to fix that.

Now, I've designed and written a basic mock auton routine for the left
side of the field.
In that, the functions and route itself are already planned and
written-- just that the exact coordinates, angles and timings need to
be actually recorded and tested.

It's a pretty simple but ambitious routine and demonstrates the strengths
of pure-pursuit as otherwise I could've just used a PID and that thought
makes me feel sad.
I first score the preload into the long goal, then go for the 3 blocks
in the top corner of the field and score them in the middle goal before
finding a nice final place.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>