minor typo bugfixes
Changes made:
  * pure_pursuit.zig:
    * (#) chose last_end instead of true path end
    * (#) ldr & rdr were swapped for left & right drives.

I had used my code to benchmark the reasoning skills of ChatGPT 5.
And, incidentally, and competely accidentally, it caught some subtle but
obvious typo-kind of errors.

The last_end one is a little more subtle, as at the start the last_end
is initialized to 0, so it would compare against the start of the path.
So if the robot was within precision of the start of the path, it would
exit pre-maturely.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>