physical robot configs & bugfixes
Changes made:
  * drive.zig:
    * (~) updated the drivetrain motor ports
    * (#) split arcade joysticks were swapped
  * tower.zig:
    * (+) added scoring middle again
    * (~) updated tower motor port & comments
    * (#) added pneumatics solenoid configuration in init
    * (~) split little-will controls from tower controls if chain
  * odom.zig: updated imu port

Today we've added cables to the robot.
And, so, I've been able to not test auton, but just double check
that opcontrol is working.
We've also updated the new ports (no odom ports or configs yet).

Through this, we've discovered that we can score middle again, and
also a bunch of bugs that have been promptly fixed.

We'll test auton at comp :sob:

Signed-off-by: Ethan Ma <kalscium@protonmail.com>