competition commit
Changes Made:
  * autonomous.zig:
    * (~) further tuned/modified the auton pure pursuit params
    * (+) added a testing auton path without pure pursuit
    * (~) added autonLeft's actual recorded coordinates (for pp)
  * drive.zig:
    * (~) changed the reverse toggle button to be the down arrow
    * (~) split the movement speed into a separate driving & turning speed
    * (~) changed split arcade's right joystick to control x instead of y
    * (~) changed toggle-arcade toggles to be Y & A instead of R1 & R2
    * (~) removed the locking of axises for y movement (toggle arcade)
  * (~) logging.zig: updated tower motor names
  * odom.zig:
    * (~) changed position recording button to be the up arrow
    * (~) updated the rotation sensor ports & configs
    * (!) note: lateral odom does NOT WORK
  * pid.zig:
    * (#) forgot to use the absolute value of error when checking for precision thresholds (rotation)
    * (#) accidentally used yaw instead of error for the rotation pid update
  * tower.zig:
    * (~) updated controls and speeds according to driver preference
    * (~) renamed tower motors to make more sense (top & bottom)
    * (~) split tower & little-will controls

Yesterday we had our second robotics competition at St Andrews.
Which also entailed, the first time I would get to properly test
autonomous since last month, and also the first time the drivers would
get to drive the robot.

So, after much trial and error, and communication with the drivers,
I've updated a bunch of the controls for the robot drive & towers to
fit their preference. Also, the physical requirements of the robot has
also changed, so I've had to account for those changes too.

We ended up coming last in the comp due to our autonomous not working
due to a lack of testing, and also our scoring/storing of blocks and
little will mech being inadequate for the exact same reasons.

At the competition was also when I was able to test autonomous, and
record the correct coordinates for the path (as you can see in the diff).
Though, despite the path working fine at the correct speed during
debug-mode, once it was run during the comp (autonomous period) the
auton ran at a speed FAR too high and just didn't work
(we need more testing).

Also, eventually, due to pure pursuit feeling too inaccurate, I tried
just doing the conventional PID autonomous path.
Then, that led me to discover a mutlitude of bugs that were hiding in the
rotate pid code which we should've fixed.

Also, the lateral odom code does not work in that the calculations are
off and lead the robot to spiral for some reason.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>